pyproject.toml
[tool.poetry]
name = "nonebot-plugin-bertvits2"
version = "0.1.0"
description = ""
authors = ["Sora <atri@suzakuintsubaki.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nonebot2 = "^2.1.1"
aiohttp = "^3.8.6"
nonebot-adapter-onebot = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
